Changes after review1

+++ day4/serviceStation/Person.java
@@ -7,10 +7,10 @@ import java.io.*;
 class Person{
 	private String name;
 	private int age;
-	private int contact;
+	private long contact;
 	private int empId;
 
-	public void setValues(String name, int age, int contact, int empId){
+	public void setValues(String name, int age, long contact, int empId){
 		this.name=name;
 		this.age= age;
 		this.contact= contact;
